requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==2.1.1
contourpy==1.0.7
cycler==0.11.0
dateparser==1.1.7
fonttools==4.38.0
frozenlist==1.3.3
gevent==22.10.2
greenlet==2.0.2
idna==3.4
joblib==1.2.0
kiwisolver==1.4.4
matplotlib==3.6.3
multidict==6.0.4
numpy==1.24.2
packaging==23.0
pandas==1.5.3
patsy==0.5.3
Pillow==9.4.0
pyparsing==3.0.9
python-binance==1.0.16
python-dateutil==2.8.2
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
regex==2022.10.31
requests==2.28.2
scikit-learn==1.2.1
scipy==1.10.0
six==1.16.0
statsmodels==0.13.5
threadpoolctl==3.1.0
tzdata==2022.7
tzlocal==4.2
ujson==5.7.0
urllib3==1.26.14
websocket-client==1.5.1
websockets==10.4
yarl==1.8.2
zope.event==4.6
zope.interface==5.5.2
